+++ src/main/java/com/example/msgadminapi/domain/repository/DayOfWeekRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DayOfWeekRepository extends JpaRepository<DayOfWeek, Long> {
+public interface DayOfWeekRepository extends JpaRepository<DayOfWeek, Integer> {
 }
